requirements.txt
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.5
zope.interface==5.4.0
